apps/illustration-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.0
Pillow==10.4.0
numpy==1.26.4
replicate==0.32.0
python-multipart==0.0.9
